document_ai_warehouse/batch-ingestion-docai-wh/requirements.txt
google-cloud-logging
google-cloud-documentai==2.16.1
google-cloud-storage
google-cloud-contentwarehouse==0.5.1
fireo==1.4.1
google-cloud-bigquery==2.20.0
pandas